+++ FractionateEverything/CHANGELOG.md
@@ -10,6 +10,7 @@
     + 调整部分矿物复制配方的概率，添加黑雾独有材料的矿物复制配方。
     + 增加前台自动百连抽功能。
     + 重新调整配方效果曲线，现在品质突破不会降低配方效果。
+    + 优化分馏塔的处理逻辑。
     + The statistics panel can correctly display the consumption of each essence during quantum copying.
     + Add the metadata needed to unlock all technologies.
     + Repair the number of matrices required for certain technologies.
@@ -24,6 +25,7 @@
       unique to Black Mist.
     + Added automatic 100 consecutive draws feature to the front desk.
     + The recipe effect curve has been readjusted, and now quality breakthroughs will not reduce the recipe effect.
+    + Optimise the processing logic of the fractionator.
 
 - v2.0.4
     + 分馏塔提示窗口内容受配方信息显示情况控制。